0.0.1: Base supports
At this time the client can estabilish a connection with server, get a
session id and request valued variables e objects as reference. Can get
and set properties of .NET objects and call his method. A base
management of exceptions is present.

For execute test use npm "test" (for windows) or "testlinux" scripts